Add the ability to add comments
Some of the views and the functionality in the comments controller isn't
currently used but might be in the future so I've left it in, for example the
ability to edit and delete comments.

The comments are nested inside Item objects, and redirects in the controller go
back to the Item the comment belongs to.