Exclude unwanted content from search
At the moment the remote content we get back from HTML pages includes tags such
as `<style>`. We don't want the content of these tags to appear in the search,
so we're excluding a list to remove from the content. We can add more tags to
this list if we find more to exclude over time.